Move ThreadInfo to services/timetable
